Split SSEEventDispatcher into its own module to drop inline imports
SSEEventDispatcher lived in galaxy.managers.sse alongside the pure connection
types (SSEEvent, SSEConnectionManager). That forced two local imports inside
_send() — send_control_task and all_control_queues_for_declare — because
galaxy.queue_worker depends on SSEConnectionManager/SSEEvent from sse.py, so
hoisting the queue_worker import to the top of sse.py would create a cycle.

Move the dispatcher to galaxy.managers.sse_dispatch. It can then import from
galaxy.queue_worker and galaxy.queues at module top: the new module is a leaf
with respect to those dependencies (queue_worker never imports it back), so
there's no cycle. Callers (app, notification manager, history audit monitor)
import SSEEventDispatcher from its new home.

Note: the conditional `import psycopg` / `import psycopg2` in
history_audit_monitor.py and the feature-flagged HistoryAuditMonitor import
in app.py are retained — both are legitimate optional/gated loads, not
cycle workarounds.
